feat(*): update the comments and refine the code style
1) comments; 2) struct;
