+++ external/network.go
@@ -4,16 +4,12 @@ import "github.com/Grivn/phalanx/common/protos"
 
 // NetworkService has provided the service for network communication.
 type NetworkService interface {
+	// BroadcastCommand is used to send the commands generated by proposer nodes.
+	BroadcastCommand(command *protos.Command)
+
 	// BroadcastPCM is used to send the message to every node in current cluster.
-	// in this implement of phalanx, we prefer a PhalanxBroadcast in which we could send message to ourself.
 	BroadcastPCM(message *protos.ConsensusMessage)
 
 	// UnicastPCM is used to send the message to the target node.
 	UnicastPCM(message *protos.ConsensusMessage)
-
-	TestSender
-}
-
-type TestSender interface {
-	BroadcastCommand(command *protos.Command)
 }
